Add Treasuries and it's moves Change how the StartUp_Engine works to be Async Change how and where logs saved New InitialModel Change Model id from int to long remove dates from moves Replace the A7 Bill Report with A8 Replace C2 string format with N in user input fields

+++ src/Model/Bill.cs
@@ -10,11 +10,11 @@ namespace Phony.Model
             BillsServicesMoves = new ObservableCollection<BillServiceMove>();
         }
 
-        public int ClientId { get; set; }
+        public long ClientId { get; set; }
 
         public virtual Client Client { get; set; }
 
-        public int StoreId { get; set; }
+        public long StoreId { get; set; }
 
         public virtual Store Store { get; set; }
